Revert "Merge origin/master into multichannelGrabber"
This reverts commit 837d5b0d3ef5101ee3101d7d2d013155e9dfedc5, reversing
changes made to 755d1ac4b8f9790c2d2a13bf785f6beab26c4d7e.

+++ demo/itom/pointClouds/demo_PointCloud.py
@@ -5,8 +5,11 @@ This demo is a very short introduction to basic point clouds
 and a polygon mesh, that consists of two triangles, whose corner
 points are given by some points of a cloud."""
 
+
 import numpy as np
-from itom import point, pointCloud, polygonMesh
+from itom import pointCloud
+from itom import polygonMesh
+from itom import point
 
 # sphinx_gallery_thumbnail_path = '11_demos/_static/_thumb/demoPCL.png'
 
@@ -50,12 +53,12 @@ for idx in range(pcl.size):
 # to the points in the cloud that define a triangle or quadrangle of the
 # mesh. If the normal vector to such a triangle points towards the outside
 # of the object, the order of the points is given by the right-hand rule, where
-# the thumbs shows in the direction of the normal vector.
+# the thumbs shows in the direction of the normale vector.
 
 # create an array with two triangles. The first triangle has the corner
 # points with the indices 0, 1 and 2 of the cloud. The 2nd triangle consists
 # of the corner points 0, 1 and 3.
-vertices = np.array([[0, 1, 2], [0, 1, 3]]).astype("int32")
+vertices = np.array([[0, 1, 2], [0, 1, 3]])
 
 # create a mesh from the cloud and the vertices
 mesh = polygonMesh.fromCloudAndPolygons(pcl, vertices)
